+++ skyflash/utils.py
@@ -9,11 +9,16 @@ import traceback
 import subprocess
 import json
 import ipaddress
+import requests
 from PyQt5.QtCore import QObject, pyqtSignal, QRunnable, pyqtSlot
 
 if 'nt' in os.name:
     import wmi
     import win32file
+
+# version data
+actualVersion = "v0.0.4beta1"
+updateURL = "https://raw.githubusercontent.com/skycoin/skyflash/master/version.txt"
 
 def cleanString(data):
     '''Cleans a string from trailing or leading chars
@@ -600,3 +605,42 @@ class Worker(QRunnable):
         NO_ASSOC = 31
         OOM = 8
         SHARE = 26
+
+def checkUpdates(data_callback, progress_callback):
+    '''This routine compare the actual noted version against the one
+    in the official repository to check for updates
+    
+    If we can't ge the file we return null
+    If reached and the same False
+    If reached and different True
+    
+    '''
+
+    version = ""
+
+    try:
+        # try to obtain the file with the latest version
+        r = requests.get(updateURL)
+        if r.status_code != requests.codes.ok:
+            return "None"
+    except:
+        return "None"
+
+    # we get a response
+    data = r.text.splitlines()
+    for line in data:
+        if line == "":
+            continue
+
+        if line.startswith("#"):
+            continue
+
+        if line.startswith("v"):
+            version = line
+    
+    if version == actualVersion:
+        # you are in the same version
+        return "False"
+    else:
+        # yep, you have to updates
+        return "True"